terminateが呼ばれた際にalarmの再生がすでに終了していないか確認するように修正

+++ lib/alarm_player.py
@@ -19,4 +19,5 @@ class AlarmPlayer:
 
     def terminate(self):
         self.is_stopped = True
-        self.proc.terminate()
+        if self.proc.poll() is None:
+            self.proc.terminate()
